Fix warnings
Fix a runtime SyntaxWarning due to an unhandled '\e' escape sequence and
other warnings obtained when building the documentation with Sphinx.

The 'html_domain_indices' change is not involved in any warning, I'm
just restoring it to its default value which appears to be True
nowadays. I don't see any difference between True and False on this
document.